+++ Readme/CHANGELOG.md
@@ -40,6 +40,7 @@ If you would like to see fresh changes done since last release you can do so on 
 
   - Added new module `Ruleset.Compatibility` because `Appx` module no longer works since PowerShell
   Core 7.1
+  - Improved function `Approve-Execute`
 
 - Documentation